Create bundle and the first feature narrative
$> app/console generate:bundle -n --namespace Bossa/Bundle/TrainingBundle --dir src
$> mkdir -p src/Bossa/Bundle/TrainingBundle/Features/admissions
$> vim src/Bossa/Bundle/TrainingBundle/Features/admissions/student_enrols_on_a_course.feature

We don't need Behat for starting the conversation. We can jump into describing
the features. Features folder go under the root of the bundle. Inside the
Features folder we create as many folders as features we will have. The will
group the .feature files that contain the different user stories and scenarios
for that feature.