Add some scenarios
$> vim src/Bossa/Bundle/TrainingBundle/Features/admissions/student_enrols_on_a_course.feature

We are adding just enough scenarios to go over the steps of full stack bdd
with Symfony. Please note that the order of the scenarios is important. They
should ease the conversation. They should be fluid, meaning current scenario
should expand the knowledge of previous one and prepare the conversation for
the next one. They should be focused on the domain. They should focus on
value and provide the "what do we mean" for the feature narrative.