Implement first step definition
$> vim src/Bossa/Bundle/TrainingBundle/Features/Context/StudentContext.php
$> vim src/Bossa/Bundle/TrainingBundle/Features/Context/FeatureContext.php

We implement `there are no courses scheduled` step definition. But, as we
know that application state is clean at the beginning of each scenario,
we're just removing `PendingException` from step definition, marking it
as passed. Still in this case, we need to make sure, that we have clean
application state before each scenario, so we're adding `@BeforeScenario`
hook to purge database with `doctrine/data-fixtures` package.