Implement second step definition
$> vim src/Bossa/Bundle/TrainingBundle/Features/Context/StudentContext.php
$> vim app/config/routing.yml

In implementing the visiting pages step we recommend that you take advantage
of expressive route names. That forces us to add the route in the
configuration. Because no class is required, no specs are written. Making sure
you visit the url and that the route exists is not enough though. You need to
know that the status returned is 200. Which takes us to implementing the
controller that will handle that route.