Start describing controller
copy gist to spec/ControllerSpecification.php
$> bin/phpspec describe Bossa/Bundle/TrainingBundle/Controller/CoursesController
$> cat spec/Bossa/Bundle/TrainingBundle/Controller/CoursesController.php
$> bin/phpspec run
$> vim spec/Bossa/Bundle/TrainingBundle/Controller/CoursesController.php
$> bin/phpspec run
$> vim src/Bossa/Bundle/TrainingBundle/Controller/CoursesController.php

We have a failing scenario, and we feel we need to add some code. It's time
to introduce our first spec. We install phpspec and tell it we want to
describe our controller. PHPSpec will create the controller specification
and it will generates the required code for us. We will then just add the
relevant behaviour we are expecting and make it green.'