Doing couple of scenario cycles about our entity
$> vim behat.yml
We've enabled page showing in extension for more output
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
We've ran scenario and it sais "No Bossa entity namespace"
$> vim app/config/config.yml
So we're adding configuration to the config.yml
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
We're running scenario again to discover "Directory does not exist"
$> mkdir src/Bossa/Entity
So we've created it
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
And then scenario started to complain about "No class"
$> bin/phpspec describe Bossa/Entity/Course
So we've described it
$> bin/phpspec run
And checked it. PHPSpec2 was nice enough to generate class for us
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
Then Doctrine2 inside scenario, started to complain that there's no mapping
$> vim spec/Bossa/Entity/Course.php
So we've added entity mapping
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
Then, scenario said "No primary key"
$> vim spec/Bossa/Entity/Course.php
So we've added getter for property and property example into spec
$> bin/phpspec run
And PHPSpec2 again was nice enough to generate method for us.
$> vim spec/Bossa/Entity/Course.php
We just needed to add property and its mapping.
$> bin/behat --no-paths admissions/student_enrols_on_a_course.feature:6
And we checked that we have no-Doctrine2 exception now

We're going couple of scenario exception cycles to completely get rid
of Doctrine exceptions. This is called "Change the message or make it
green".